dvpingpong: Add synchronization between traffic thread and event thread
Use semaphore to synchronize the traffic and event thread:
- The event thread post the semaphore on a cq completion event;
- The traffic thread do poll_cq only when a semaphore is received, to
  avoid busy-loop.

Note that semaphore is not a good example, as it's just one signal.
If the event eq is shared by multiple CQs then the event thread needs
to notify the traffic thread with the cq number. In this case a better
solution is need, e.g., pipe(2).

Signed-off-by: Mark Zhang <markzhang@nvidia.com>